split login scripts (module issue still present)
