+++ project/nodeapp/views/hashedlogin.ejs
@@ -12,10 +12,10 @@
           <p id="login-error-msg">Invalid username <span id="error-msg-second-line">and/or password</span></p>
         </div>
           <form id="login-form">
-            <h1 id="login-header">Login</h1>
+            <h1 id="login-header">Login: Hashed edition</h1>
             <input type="password" name="password" id="password-field" class="login-form-field" placeholder="Password">
             <input type="submit" value="Login" id="hashed-login-form-submit">
           </form>
-        <script src="../login-page.js"></script>
+        <script src="../hashed-login-page.js"></script>
       </body>
 </html>
